+++ src/main/java/com/posgrado/gradosytitulos/dto/dto/students/StudentsCreate.java
@@ -17,7 +17,7 @@ public record StudentsCreate (
         @NotBlank
         String phone,
         @NotNull
-        Integer degreeId,
+        Long degreeId,
         String photo
 ) {
 }
