+++ src/main/java/com/posgrado/gradosytitulos/dto/dto/students/StudentsViewDTO.java
@@ -8,7 +8,7 @@ public record StudentsViewDTO(
         String maternalSurname,
         String email,
         String phone,
-        Integer degreeId,
+        Long degreeId,
         String photo
 ) {
 }
